Added normalize, refactored some classes
