+++ Silhouette/src/StyleSheets/Animation.java
@@ -1,6 +1,5 @@
 package StyleSheets;
 
-import StyleSheets.BaseComponents.Implementation.Statement;
 import StyleSheets.BaseComponents.AtRule;
 
 public class Animation extends AtRule {
